+++ src/sys/os/term/metric.rs
@@ -5,7 +5,10 @@
 
 #[doc = crate::_tags!(term layout ffi)]
 /// The size of the terminal.
-#[doc = crate::_doc_meta!{location("sys/os/term")}]
+#[doc = crate::_doc_meta!{
+    location("sys/os/term"),
+    test_size_of(TermSize = 8|64),
+}]
 ///
 /// ## Used by
 /// - `LinuxTermios` and `WindowsConsole`.